backend/requirements.txt
slowapi==0.1.9
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncio==4.0.0
asyncpg==0.31.0
bcrypt==4.2.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
greenlet==3.3.2
idna==3.11
passlib==1.7.4
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
alembic==1.13.1
psycopg2-binary==2.9.9
cryptography==42.0.5
python-multipart==0.0.9

uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
httpx==0.28.1
